refactor: move styles/system.ts to stitches.config.ts
