+++ site/src/styles/components.ts
@@ -1,4 +1,4 @@
-import {styled} from './system'
+import {styled} from '../stitches.config'
 
 export const Box = styled('div', {})
