+++ site/src/styles/index.ts
@@ -1,5 +1,4 @@
 export * from './components'
 export * from './global'
 export * from './Prose'
-export * from './system'
 export * from './themes'
